+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.28.6](https://github.com/houseninjadojo/app/compare/v1.28.5...v1.28.6) (2023-01-04)
+
+
+### Bug Fixes
+
+* this.cardNumber.substr not a function ([#1497](https://github.com/houseninjadojo/app/issues/1497)) ([bdd41ef](https://github.com/houseninjadojo/app/commit/bdd41efd73bbfce5d31e9fb7bb9dd1b2d432cd02))
+
+## [1.28.5](https://github.com/houseninjadojo/app/compare/v1.28.4...v1.28.5) (2022-12-13)
+
+
+### Bug Fixes
+
+* **build:** be specific about asset bundling split ([#1412](https://github.com/houseninjadojo/app/issues/1412)) ([52a57e5](https://github.com/houseninjadojo/app/commit/52a57e5540bda32862429e3aeaae7cfb92bc93bd))
+
+## [1.28.4](https://github.com/houseninjadojo/app/compare/v1.28.3...v1.28.4) (2022-12-12)
+
+
+### Bug Fixes
+
+* **data:** polyfill crypto.randomUUID ([#1410](https://github.com/houseninjadojo/app/issues/1410)) ([2b57fd4](https://github.com/houseninjadojo/app/commit/2b57fd469fdc376491d0e88c222b9ee7544d6328))
+
 ## [1.28.3](https://github.com/houseninjadojo/app/compare/v1.28.2...v1.28.3) (2022-12-02)